Create `database.ts` file
The new file handles the Prisma instance and initialization.
